requirements.txt
streamlit
openai-whisper
pyaudio 
ffmpeg-python